// apps/konquest-web/src/app/main/content-management/models/learn-content-list-item-action.ts
export enum LEARN_CONTENT_LIST_ITEM_ACTION {
  DELETE = 'DELETE',
  DUPLICATE = 'DUPLICATE',
  EDIT = 'EDIT',
  EDIT_CONTRIBUTORS = 'EDIT_CONTRIBUTORS',
  ENROLL_USERS = 'ENROLL_USERS',
  FINISH_EVENT = 'FINISH_EVENT',
  SHARE = 'SHARE',
  TRANSFER = 'TRANSFER',
  VIEW_AS_USER = 'VIEW_AS_USER',
  VIEW_EVALUATIONS = 'VIEW_EVALUATIONS',
  VIEW_STATISTICS = 'VIEW_STATISTICS',
  VINCULATE_TO_GROUP = 'VINCULATE_TO_GROUP',
  PUBLISH = 'PUBLISH',
  DETAILS = 'DETAILS',
  MANAGE_ENROLLMENTS = 'MANAGE_ENROLLMENTS',
  MANAGE_PULSES = 'MANAGE_PULSES',
}